+++ backslash/contrib/slash_plugin.py
@@ -194,6 +194,7 @@ class BackslashPlugin(PluginInterface):
         self.test_skip(reason=reason)
         self.test_end()
 
+
     @handle_exceptions
     def test_interrupt(self):
         if self.current_test is not None:
@@ -406,6 +407,20 @@ class BackslashPlugin(PluginInterface):
 
     @handle_exceptions
     def error_added(self, result, error):
+        self._add_exception(result=result, exception=error)
+
+    @slash.plugins.register_if(hasattr(slash.hooks, 'interruption_added'))
+    @handle_exceptions
+    def interruption_added(self, result, exception):
+        self._add_exception(result=result, exception=exception, is_interruption=True)
+
+
+    def _add_exception(self, result, exception, is_interruption=False):
+        has_interruptions = self.client.api.info().endpoints.add_error.version >= 4
+        if is_interruption and not has_interruptions:
+            _logger.debug('Server does not support recording is_interruption exceptions. Skipping reporting')
+            return
+
         if result is slash.session.results.global_result:
             error_container = self.session
         else:
@@ -415,15 +430,20 @@ class BackslashPlugin(PluginInterface):
             _logger.debug('Could not determine error container to report on for {}', result)
             return
 
-        kwargs = {'exception_type': error.exception_type.__name__ if error.exception_type is not None else None,
-                  'traceback': distill_slash_traceback(error), 'exception_attrs': getattr(error, 'exception_attributes', NOTHING)}
-        if error.message:
-            message = error.message
-        elif hasattr(error, 'exception_str'):
-            message = error.exception_str
+        kwargs = {'exception_type': exception.exception_type.__name__ if exception.exception_type is not None else None,
+                  'traceback': distill_slash_traceback(exception), 'exception_attrs': getattr(exception, 'exception_attributes', NOTHING)}
+        if exception.message:
+            message = exception.message
+        elif hasattr(exception, 'exception_str'):
+            message = exception.exception_str
         else:
-            message = str(error.exception)
+            message = str(exception.exception)
+
+
         kwargs['message'] = message
+
+        if has_interruptions:
+            kwargs['is_interruption'] = is_interruption
 
         for compact_variables in [False, True]:
             if compact_variables: